changes consisting of : 1. making criticle path turn off and on inturupts within its function call for clarity. 2. moved delay extension dsg stuff into its partitioned counter part. 3. added a message about where to find binary paths for end users to know in case they have binary issues
